Rename service and move to DataContainer namespace
