+++ core-bundle/config/services.yaml
@@ -187,12 +187,6 @@ services:
     contao.csrf.token_storage:
         class: Contao\CoreBundle\Csrf\MemoryTokenStorage
 
-    contao.database_util:
-        class: Contao\CoreBundle\Util\DatabaseUtil
-        public: true
-        arguments:
-            - '@database_connection'
-
     contao.data_collector:
         class: Contao\CoreBundle\DataCollector\ContaoDataCollector
         arguments:
@@ -239,6 +233,12 @@ services:
             - '@contao.translation.translator'
             - '@contao.data_container.record_labeler'
             - '@contao.data_container.dca_request_switcher'
+            - '@database_connection'
+
+    contao.data_container.foreign_key_parser:
+        class: Contao\CoreBundle\DataContainer\ForeignKeyParser
+        public: true
+        arguments:
             - '@database_connection'
 
     contao.data_container.global_operations_builder:
@@ -1532,6 +1532,7 @@ services:
     Contao\CoreBundle\Csp\WysiwygStyleProcessor: '@contao.csp.wysiwyg_style_processor'
     Contao\CoreBundle\Csrf\ContaoCsrfTokenManager: '@contao.csrf.token_manager'
     Contao\CoreBundle\DataContainer\DcaUrlAnalyzer: '@contao.data_container.dca_url_analyzer'
+    Contao\CoreBundle\DataContainer\ForeignKeyParser: '@contao.data_container.foreign_key_parser'
     Contao\CoreBundle\Doctrine\Backup\DumperInterface: '@contao.doctrine.backup.dumper'
     Contao\CoreBundle\Filesystem\FileDownloadHelper: '@contao.filesystem.file_download_helper'
     Contao\CoreBundle\Framework\ContaoFramework: '@contao.framework'
@@ -1563,5 +1564,4 @@ services:
     Contao\CoreBundle\Twig\Finder\FinderFactory: '@contao.twig.finder_factory'
     Contao\CoreBundle\Twig\Interop\ContextFactory: '@contao.twig.interop.context_factory'
     Contao\CoreBundle\Twig\Loader\ContaoFilesystemLoader: '@contao.twig.filesystem_loader'
-    Contao\CoreBundle\Util\DatabaseUtil: '@contao.database_util'
     Contao\CoreBundle\Util\ProcessUtil: '@contao.process_util'
